UI improvments, save, delete, get, checkmark functions became working.
