+++ InfinityTo-Do's/View/HomePageView/ToDoListView/NewItemView/NewItemView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct NewItemView: View {
      
-     @StateObject var viewModel = AddNewToDoListViewModel()
+     @StateObject var viewModel = NewItemViewModel()
      @Binding var newItemPresented: Bool
